Takes CI from every lane red — three of them not even building — to
everything building, Linux fully green, format green, and both sanitizer
lanes green.

- 42 -Werror sites across 16 files. The build died in econlife_core, so
  40 of them were unreachable by the compiler and CI could only ever
  report two. Every numeric fix is an explicit cast where the compiler
  already performed that conversion implicitly.
- Tree-wide clang-format (59 files had drifted).
- Linux Debug link parallelism capped at -j 2: the runner was being
  OOM-killed during linking.
- Windows: three unit tests relied on a transitive <unistd.h> for
  getpid(), which MSVC does not have. tests/test_process_id.h now
  provides it portably.
- A real data race on EnergyBaseModule::energy_state_dirty_, written
  from execute_province() — one thread per province. The write moves to
  init_for_tick(), on the main thread; the predicate is unchanged.
- The long integration tests carry their own TIMEOUT, so one number
  covers Debug, Release, Windows and both sanitizer lanes instead of
  four lane defaults disagreeing.

Restoring the Windows build exposed 16 pre-existing behavioural failures
there (the identical commit passes 1701/1701 on Linux). They are not
caused by this change and are recorded in docs/session_logs/flagged_issues.md
for separate work; the cross-platform determinism guarantee went
unverified on half the matrix for as long as that build was broken.

Also records that neither EnergyBaseModule::execute() nor
SubsistenceModule::execute() is ever called — both are province-parallel
and neither overrides has_global_post_pass() — so their regime-exit
resets have never run. Fixing that activates dormant behaviour in the
demographic path and is deliberately left as a design decision.

+++ simulation/modules/grain_logistics/grain_logistics_module.cpp
@@ -73,7 +73,8 @@ void GrainLogisticsModule::execute(const WorldState& state, DeltaBuffer& delta) 
             continue;
 
         for (const auto& d : dests) {
-            const double alloc = static_cast<double>(surplus) * (static_cast<double>(d.second) / total_w);
+            const double alloc =
+                static_cast<double>(surplus) * (static_cast<double>(d.second) / total_w);
             const double delivered = alloc * static_cast<double>(d.second);  // ox eats (1 - df)
             net_feedable[d.first] += delivered;
         }
@@ -96,8 +97,9 @@ void GrainLogisticsModule::execute(const WorldState& state, DeltaBuffer& delta) 
         avail[i] = std::max(0.0f, cs.food_store);
         // The province's own reserve target: what a full granary means here.
         target[i] = static_cast<double>(cs.total_population) *
-                    subsistence_cfg_.per_capita_food_per_tick * tpy *
-                    subsistence_cfg_.granary_reserve_years;
+                    static_cast<double>(subsistence_cfg_.per_capita_food_per_tick) *
+                    static_cast<double>(tpy) *
+                    static_cast<double>(subsistence_cfg_.granary_reserve_years);
     }
     if (rate_tick > 0.0f) {
         for (uint32_t src = 0; src < n; ++src) {
@@ -122,11 +124,12 @@ void GrainLogisticsModule::execute(const WorldState& state, DeltaBuffer& delta) 
                 const double f_dst = avail[dst] / target[dst];
                 if (f_src <= f_dst)
                     continue;  // grain flows down the scarcity gradient only
-                const double sent =
-                    std::min(avail[src], rate_tick * (f_src - f_dst) * target[dst]);
+                const double sent = std::min(
+                    avail[src], static_cast<double>(rate_tick) * (f_src - f_dst) * target[dst]);
                 if (sent <= 0.0)
                     continue;
-                const double delivered = sent * df;  // the oxen eat the rest en route
+                const double delivered =
+                    sent * static_cast<double>(df);  // the oxen eat the rest en route
                 avail[src] -= sent;
                 store_delta[src] -= sent;
                 avail[dst] += delivered;
@@ -147,7 +150,7 @@ void GrainLogisticsModule::execute(const WorldState& state, DeltaBuffer& delta) 
         RegionDelta rd{};
         rd.region_id = state.provinces[i].region_id;
         rd.net_feedable_surplus_replacement = static_cast<float>(net_feedable[i]);
-        float urban = static_cast<float>(net_feedable[i] / per_capita);
+        float urban = static_cast<float>(net_feedable[i] / static_cast<double>(per_capita));
         if (state.provinces[i].cohort_stats) {
             const float pop = static_cast<float>(state.provinces[i].cohort_stats->total_population);
             if (urban > pop)
